upvote and downvote

+++ pages/api/editArticle.js
@@ -13,29 +13,31 @@ export default async function handler(req, res) {
                 let article = await Article.findOne({
                     url: req.body.url
                 })
-                if (req.body.delete == "true") {
-                    await article.delete()
-                    res.status(200).json({
-                        success: "Article deleted successfully"
-                    })
-                } else {
-                    article.update({
-                        title: req.body.title,
-                        url: req.body.url,
-                        imageUrl: req.body.imageUrl,
-                        description: req.body.description,
-                        longDescription: req.body.content
-                    }, function (err) {
-                        if (err) {
-                            res.status(400).json({
-                                error: err
-                            })
-                        } else {
-                            res.status(200).json({
-                                success: "Article modified successfully"
-                            })
-                        }
-                    })
+                if (article.createdBy.equals(user._id) || user.role === "admin"){
+                    if (req.body.delete == "true") {
+                        await article.delete()
+                        res.status(200).json({
+                            success: "Article deleted successfully"
+                        })
+                    } else {
+                        article.update({
+                            title: req.body.title,
+                            url: req.body.url,
+                            imageUrl: req.body.imageUrl,
+                            description: req.body.description,
+                            longDescription: req.body.content
+                        }, function (err) {
+                            if (err) {
+                                res.status(400).json({
+                                    error: err
+                                })
+                            } else {
+                                res.status(200).json({
+                                    success: "Article modified successfully"
+                                })
+                            }
+                        })
+                    }
                 }
             }
         }
